backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.4
PyJWT==2.10.1
argon2-cffi==23.1.0

# External APIs
httpx==0.27.2
google-api-python-client==2.161.0
google-auth==2.38.0

